FIX: 라벨만 추가되면 구현 가능

+++ backend/src/main/java/org/mint/smallcloud/file/service/DirectoryFacadeService.java
@@ -2,6 +2,8 @@ package org.mint.smallcloud.file.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.mint.smallcloud.exception.ExceptionStatus;
+import org.mint.smallcloud.exception.ServiceException;
 import org.mint.smallcloud.file.domain.File;
 import org.mint.smallcloud.file.domain.Folder;
 import org.mint.smallcloud.file.dto.*;
@@ -68,6 +70,8 @@ public class DirectoryFacadeService {
 
     public void purge(Long directoryId, String username) {
         Folder folder = directoryThrowerService.getDirectoryById(directoryId);
+        if (folder.isRoot())
+            throw new ServiceException(ExceptionStatus.NO_PERMISSION);
         directoryThrowerService.checkAccessRight(folder, username);
         directoryService.purgeDirectory(folder);
     }
@@ -76,5 +80,11 @@ public class DirectoryFacadeService {
         Folder folder = directoryThrowerService.getDirectoryById(directoryId);
         directoryThrowerService.checkAccessRight(folder, username);
         directoryService.deleteDirectory(folder);
+    }
+
+    public void restore(Long directoryId, String username) {
+        Folder folder = directoryThrowerService.getDirectoryById(directoryId);
+        directoryThrowerService.checkAccessRight(folder, username);
+        directoryService.restoreDirectory(folder);
     }
 }
